LayoutTests/platform/chromium-win/media/video-src-expected.txt

EVENT(loadstart)
EXPECTED (relativeURL(video.currentSrc) == 'content/test.ogv') OK
END OF TEST
